Revert "Made filetype matching a bit more flexible"
This reverts commit 289d104758fa3a95fdba201832510d5a7c81c8d9, merged to
https://github.com/martinda/Jenkinsfile-vim-syntax in PR #12.

Unfortunately, PR #12 makes filetype matching a bit too flexible.  It
broke filetype detection for non-Jenkinsfile files whose names contain
the string "jenkins".  For example, it forced the filetype of
`ftdetect/Jenkinsfile.vim` to be detected as `Jenkinsfile` instead of
`vim`, which is bad.

Vim already includes multiple mechanisms to detect nonstandard filenames
as Jenkinsfiles:

- Add a custom `autocmd` group to the local `.vimrc`.
- Add a Vim modeline to the file itself.

Updating a public plugin to accommodate local nonstandard behavior is
problematic, as it breaks functionality for all plugin users.